+++ addon/components/olv-toolbar.js
@@ -37,6 +37,14 @@ export default FlexberryBaseComponent.extend({
   objectlistviewEventsService: Ember.inject.service('objectlistview-events'),
 
   /**
+    Service for managing advLimits for lists.
+
+    @property advLimit
+    @type AdvLimitService
+  */
+  advLimit: Ember.inject.service(),
+
+  /**
     Flag to use creation button at toolbar.
 
     @property createNewButton
@@ -81,6 +89,16 @@ export default FlexberryBaseComponent.extend({
     @readOnly
   */
   colsConfigButton: true,
+
+  /**
+    Flag to use advLimitButton button at toolbar.
+
+    @property advLimitButton
+    @type Boolean
+    @default true
+    @readOnly
+  */
+  advLimitButton: false,
 
   /**
     Flag indicates whether to show exportExcelButton button at toolbar.
@@ -203,16 +221,28 @@ export default FlexberryBaseComponent.extend({
   }),
 
   /**
+    Current adv limits.
+
+    @property namedAdvLimits
+    @type Object
+  */
+  namedAdvLimits: undefined,
+
+  /**
     @property colsConfigMenu
     @type Service
   */
   colsConfigMenu: Ember.inject.service(),
 
-  menus: [
+  /**
+    @property menus
+    @readOnly
+  */
+  menus: Ember.computed(() => Ember.A([
     { name: 'use', icon: 'checkmark box' },
     { name: 'edit', icon: 'setting' },
     { name: 'remove', icon: 'remove' }
-  ],
+  ])).readOnly(),
 
   /**
     @property colsSettingsItems
@@ -275,10 +305,56 @@ export default FlexberryBaseComponent.extend({
   ),
 
   /**
-    Observe colsSettingsItems changes.
-    @property _colsSettingsItems
+    @property advLimitItems
     @readOnly
   */
+  advLimitItems: Ember.computed('i18n.locale', 'advLimit.isAdvLimitServiceEnabled', 'namedAdvLimits', function() {
+    const i18n = this.get('i18n');
+    const rootItem = {
+      icon: 'dropdown icon',
+      iconAlignment: 'right',
+      title: '',
+      items: Ember.A(),
+      localeKey: ''
+    };
+    const createLimitItem = {
+      icon: 'flask icon',
+      iconAlignment: 'left',
+      title: i18n.t('components.olv-toolbar.create-limit-title'),
+      localeKey: 'components.olv-toolbar.create-limit-title'
+    };
+    rootItem.items.addObject(createLimitItem);
+
+    const limitItems = this.get('namedAdvLimits');
+    const menus = this.get('menus');
+    const editMenus = Ember.A();
+    menus.forEach(menu => {
+      const menuSubitem = this._createMenuSubitems(limitItems, menu.icon + ' icon');
+      if (menuSubitem.length > 0) {
+        editMenus.addObject({
+          icon: 'angle right icon',
+          iconAlignment: 'right',
+          localeKey: `components.olv-toolbar.${menu.name}-limit-title`,
+          items: menuSubitem
+        });
+      }
+    }, this);
+
+    if (editMenus.length > 0) {
+      rootItem.items.addObjects(editMenus);
+    }
+
+    const setDefaultItem = {
+      icon: 'remove circle icon',
+      iconAlignment: 'left',
+      title: i18n.t('components.olv-toolbar.set-default-limit-title'),
+      localeKey: 'components.olv-toolbar.set-default-limit-title'
+    };
+    rootItem.items.addObject(setDefaultItem);
+
+    return this.get('advLimit.isAdvLimitServiceEnabled') ? Ember.A([rootItem]) : Ember.A();
+  }),
+
   _colsSettingsItems: Ember.observer('colsSettingsItems', function() {
     this._updateListNamedUserSettings(this.get('componentName'));
   }),
@@ -541,6 +617,17 @@ export default FlexberryBaseComponent.extend({
     },
 
     /**
+      Action to show confis dialog.
+
+      @method actions.showConfigDialog
+      @public
+    */
+    showAdvLimitDialog(settingName) {
+      Ember.assert('showAdvLimitDialog:: componentName is not defined in flexberry-objectlistview component', this.componentName);
+      this.get('modelController').send('showAdvLimitDialog', this.componentName, settingName);
+    },
+
+    /**
       Action to show export dialog.
 
       @method actions.showExportDialog
@@ -630,6 +717,59 @@ export default FlexberryBaseComponent.extend({
     },
 
     /**
+      Handler click on flexberry-menu of advLimits.
+
+      @method actions.onLimitMenuItemClick
+      @public
+      @param {jQuery.Event} e jQuery.Event by click on menu item
+    */
+    onLimitMenuItemClick(e) {
+      const iTags = Ember.$(e.currentTarget).find('i');
+      const namedLimitSpans = Ember.$(e.currentTarget).find('span');
+      if (iTags.length <= 0 || namedLimitSpans.length <= 0) {
+        return;
+      }
+
+      const className = iTags.get(0).className;
+      const advLimitName = namedLimitSpans.get(0).innerText;
+      const componentName = this.get('componentName');
+      const advLimitService = this.get('advLimit');
+
+      switch (className) {
+        case 'flask icon':
+          this.send('showAdvLimitDialog');
+          break;
+        case 'checkmark box icon':
+          const advLimit = this.get(`namedAdvLimits.${advLimitName}`);
+          advLimitService.saveAdvLimit(advLimit, componentName).
+            then(() => {
+              this.send('refresh');
+            });
+          break;
+        case 'setting icon':
+          this.send('showAdvLimitDialog', advLimitName);
+          break;
+        case 'remove icon':
+          advLimitService.deleteAdvLimit(componentName, advLimitName)
+          .then(() => {
+            this.get('colsConfigMenu').updateNamedAdvLimitTrigger(componentName);
+            alert(
+              this.get('i18n').t('components.advlimit-dialog-content.limit') +
+              '"' + advLimitName + '"' +
+              this.get('i18n').t('components.advlimit-dialog-content.is-deleted')
+            );
+          });
+          break;
+        case 'remove circle icon':
+          advLimitService.saveAdvLimit('', componentName)
+          .then(() => {
+            this.send('refresh');
+          });
+          break;
+      }
+    },
+
+    /**
       Handler click on flexberry-menu.
 
       @method actions.onExportMenuItemClick
@@ -697,6 +837,7 @@ export default FlexberryBaseComponent.extend({
     this.get('colsConfigMenu').on('updateNamedSetting', this, this._updateListNamedUserSettings);
     this.get('colsConfigMenu').on('addNamedSetting', this, this._addNamedSetting);
     this.get('colsConfigMenu').on('deleteNamedSetting', this, this._deleteNamedSetting);
+    this.get('colsConfigMenu').on('updateNamedAdvLimit', this, this._updateNamedAdvLimits);
   },
 
   didInsertElement() {
@@ -726,6 +867,7 @@ export default FlexberryBaseComponent.extend({
     this.get('colsConfigMenu').off('updateNamedSetting', this, this._updateListNamedUserSettings);
     this.get('colsConfigMenu').off('addNamedSetting', this, this._addNamedSetting);
     this.get('colsConfigMenu').off('deleteNamedSetting', this, this._deleteNamedSetting);
+    this.get('colsConfigMenu').off('updateNamedAdvLimit', this, this._updateNamedAdvLimits);
     this._super(...arguments);
   },
 
@@ -837,5 +979,39 @@ export default FlexberryBaseComponent.extend({
         this.get('colsSettingsItems')[0].items[i + 1].items.sort((a, b) => a.title > b.title);
       }
     }
+  },
+
+  /**
+    Refresh current adv limits list.
+
+    @method _updateNamedAdvLimits
+
+    @param {String} componentName The name of objectlistview component
+  */
+  _updateNamedAdvLimits(componentName) {
+    const advLimitService = this.get('advLimit');
+    const thisComponentName = this.get('componentName');
+    if (!(advLimitService.get('isAdvLimitServiceEnabled') && componentName === thisComponentName)) {
+      return;
+    }
+
+    this.set('namedAdvLimits', advLimitService.getNamedAdvLimits(thisComponentName));
+  },
+
+  /**
+    Creating menu subitems.
+
+    @method _createMenuSubitems
+
+    @param {Object} itemsNameList Object with items names as keys.
+    @param {String} icon Icon class for menu items.
+  */
+  _createMenuSubitems(itemsNameList, icon) {
+    if (Ember.isNone(itemsNameList)) {
+      return Ember.A();
+    }
+
+    const itemsNames = Ember.A(Object.keys(itemsNameList)).sortBy('name');
+    return itemsNames.map(name => { return { title: name, icon: icon, iconAlignment: 'left' }; });
   }
 });