+++ addon/locales/en/translations.js
@@ -165,6 +165,11 @@ export default {
       'edit-setting-title': 'Edit',
       'remove-setting-title': 'Remove',
       'set-default-setting-title': 'Default setting',
+      'create-limit-title': 'New limit',
+      'use-limit-title': 'Use',
+      'edit-limit-title': 'Edit',
+      'remove-limit-title': 'Remove',
+      'set-default-limit-title': 'Default limit',
       'show-default-setting-title': 'Show settings',
       'show-setting-caption': 'To set this settings by default you can copy and init property developerUserSettings in /app/routes/',
       'close': 'Close',
@@ -209,6 +214,21 @@ export default {
       'per-page': 'Records count on page',
       'det-separate-rows': 'List\'s properties in separate rows',
       'det-separate-cols': 'List\'s properties in separate columns'
+    },
+
+    'advlimit-dialog-content': {
+      'title': 'Customize limit',
+      'limit-name': 'Limit name',
+      'enter-limit-name': 'Enter limit name',
+      'use': 'Use',
+      'save': 'Save',
+      'have-errors': 'When you save a limit errors occurred: ',
+      'cant-parse': 'Can\'t convert current limit string to predicate',
+      'limit': 'Limit ',
+      'is-saved': ' was saved',
+      'is-deleted': ' was deleted',
+      'is-correct': 'Current limit string is correct',
+      'check': 'Check'
     },
 
     'form-load-time-tracker': {