+++ addon/routes/list-form.js
@@ -85,6 +85,14 @@ ErrorableRouteMixin, {
   appState: Ember.inject.service(),
 
   /**
+    Service for managing advLimits for lists.
+
+    @property advLimit
+    @type AdvLimitService
+  */
+  advLimit: Ember.inject.service(),
+
+  /**
     A hook you can implement to convert the URL into the model for this route.
     [More info](http://emberjs.com/api/classes/Ember.Route.html#method_model).
 
@@ -108,6 +116,8 @@ ErrorableRouteMixin, {
       this.objectListViewLimitPredicate({ modelName: modelName, projectionName: projectionName, params: params });
     let userSettingsService = this.get('userSettingsService');
     userSettingsService.setCurrentWebPage(webPage);
+    let advLimitService = this.get('advLimit');
+    advLimitService.setCurrentAppPage(webPage);
     let developerUserSettings = this.get('developerUserSettings');
     Ember.assert('Property developerUserSettings is not defined in /app/routes/' + transition.targetName + '.js', developerUserSettings);
 
@@ -138,8 +148,8 @@ ErrorableRouteMixin, {
     let userSettingPromise = userSettingsService.setDeveloperUserSettings(developerUserSettings);
     let listComponentNames = userSettingsService.getListComponentNames();
     componentName = listComponentNames[0];
-    userSettingPromise
-      .then(currectPageUserSettings => {
+    Ember.RSVP.all([userSettingPromise, advLimitService.getAdvLimitsFromStore(Object.keys(developerUserSettings))])
+      .then(() => {
         if (this._invalidSorting(params.sort)) {
           controller.set('isSortingError', true);
           transition.abort();
@@ -171,7 +181,10 @@ ErrorableRouteMixin, {
           }
         }
 
+        const advLimit = advLimitService.getCurrentAdvLimit(componentName);
+
         let queryParameters = {
+          componentName: componentName,
           modelName: modelName,
           projectionName: projectionName,
           perPage: this.perPage,
@@ -181,11 +194,13 @@ ErrorableRouteMixin, {
           filterCondition: controller.get('filterCondition'),
           filters: filtersPredicate,
           predicate: limitPredicate,
+          advLimit: advLimit,
           hierarchicalAttribute: hierarchicalAttribute,
         };
 
         this.onModelLoadingStarted(queryParameters, transition);
         this.get('colsConfigMenu').updateNamedSettingTrigger(componentName);
+        this.get('colsConfigMenu').updateNamedAdvLimitTrigger(componentName);
 
         // Find by query is always fetching.
         // TODO: support getting from cache with "store.all->filterByProjection".